+++ lib/src/widgets/widgets.dart
@@ -1,17 +1,21 @@
-export 'action_row.dart';
-export 'combo_row.dart';
-export 'clamp.dart';
-export 'flap.dart';
-export 'header_bar_minimal.dart';
-export 'header_bar.dart';
-export 'header_button.dart';
-export 'popup_menu.dart';
-export 'sidebar.dart';
-export 'stack_sidebar.dart';
-export 'preferences_group.dart';
-export 'view_switcher.dart';
-export 'view_switcher_tab.dart';
-export 'text_field.dart';
-export 'text_button.dart';
-export 'scaffold.dart';
-export 'view_stack.dart';
+// All the widgets imported from libadwaita
+export 'adw/action_row.dart';
+export 'adw/combo_row.dart';
+export 'adw/clamp.dart';
+export 'adw/flap.dart';
+export 'adw/header_bar.dart';
+export 'adw/preferences_group.dart';
+export 'adw/view_switcher.dart';
+export 'adw/view_switcher_tab.dart';
+export 'adw/view_stack.dart';
+
+// New Widgets that are not present in libadwaita
+export 'adw/new/text_field.dart';
+export 'adw/new/header_button.dart';
+export 'adw/new/scaffold.dart';
+export 'adw/new/sidebar.dart';
+export 'adw/new/text_button.dart';
+
+// All the widgets imported from Gtk4
+export 'gtk/popup_menu.dart';
+export 'gtk/stack_sidebar.dart';