+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0-rc.2
+
+- Add `AdwComboRow`
+- Increase default Header size
+- Update Sidebar theming to look more closer to libadwaita one
+- Deprecate `AdwHeaderBarMinimal`, Now use `AdwHeaderBar.minimal`
+- Reorganized widgets
+
 ## 1.0.0-rc.1
 
 - Added the following widgets: